+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Current
 
 ### Changed:
 
+- [Error messages generated during response processing include the request id.](https://github.com/yahoo/fili/pull/78)
+
 - [`DimensionStoreKeyUtils` now supports case sensitive row and column keys](https://github.com/yahoo/fili/pull/90)
     - Wrapped this config in a feature flag `case_sensitive_keys_enabled` which is set to `false` by default for backwards compatibility. 
         This flag will be set to `true` in future versions.